backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chromadb>=1.4.0",
    "fastapi>=0.127.0",
    "gitpython>=3.1.45",
    "openai>=2.14.0",
    "python-dotenv>=1.2.1",
    "tiktoken>=0.12.0",
    "uvicorn>=0.40.0",
]
